feat(tests): update comet tests to use single fixture
Update broker/archiver invocation to reflect updated cli
