+++ tests/test_vissharedmem_validation.py
@@ -1,16 +1,10 @@
 import copy
 import logging
-import numpy as np
 import os
 import pytest
-import re
-import signal
-from subprocess import Popen
-import shutil
 import tempfile
 import threading
 from time import sleep
-import sys
 
 from kotekan import runner, testing
 
@@ -19,47 +13,6 @@ useless_file = tempfile.NamedTemporaryFile()
 fname_buf = "calBuffer_" + os.path.split(useless_file.name)[-1]
 
 logging.basicConfig(level=logging.INFO)
-
-
-@pytest.fixture()
-def comet_broker():
-    broker_path = shutil.which("comet")
-    if not broker_path:
-        pytest.skip(
-            "Make sure PYTHONPATH is set to where the comet dataset broker is installed."
-        )
-
-    with tempfile.NamedTemporaryFile(mode="w") as f_out:
-        # Start comet with a random port
-        broker = Popen(
-            [broker_path, "-p", "0", "--recover", "False"], stdout=f_out, stderr=f_out
-        )
-        sleep(3)
-
-        # Find port in the log
-        regex = re.compile("Selected random port: ([0-9]+)$")
-        log = open(f_out.name, "r").read().split("\n")
-        port = None
-        for line in log:
-            print(line)
-            match = regex.search(line)
-            if match:
-                port = match.group(1)
-                print("Test found comet port in log: %s" % port)
-                break
-        if not match:
-            print("Could not find comet port in logs.")
-            exit(1)
-
-        try:
-            yield port
-        finally:
-            pid = broker.pid
-            os.kill(pid, signal.SIGTERM)
-            broker.terminate()
-            log = open(f_out.name, "r").read().split("\n")
-            for line in log:
-                print(line)
 
 
 # number of frames to ignore in validation
@@ -84,14 +37,14 @@ params_writer_stage = {"num_samples": 3, "name": fname_buf}
 
 
 @pytest.fixture()
-def vis_data(tmpdir_factory, comet_broker):
+def vis_data(tmpdir_factory, comet_broker_port):
 
     # keeping all the data this test produced here (probably do not need it)
     # using FakeVisBuffer to produce fake data
     fakevis_buffer = runner.FakeVisBuffer(**params_fakevis)
 
     # pass comet port to kotekan
-    params["dataset_manager"]["ds_broker_port"] = comet_broker
+    params["dataset_manager"]["ds_broker_port"] = comet_broker_port
 
     # KotekanStageTester is used to run kotekan with my config
     test = runner.KotekanStageTester(
@@ -106,12 +59,8 @@ def vis_data(tmpdir_factory, comet_broker):
 
 # This test still needs to run alone, because multiple comet instances would have conflicts
 # accessing redis.
-@pytest.mark.skipif(
-    sys.version_info > (3, 7),
-    reason="Test only passes when comet works, which needs old python3.7.",
-)
 @pytest.mark.serial
-def test_shared_mem_buffer(vis_data, comet_broker):
+def test_shared_mem_buffer(vis_data):
     # start kotekan writer in a thread, to read before it's done (it will delete the shm on exit)
     threading.Thread(target=vis_data.run).start()
     sleep(2)
